Renamed map storybook example. Excluded entire Storybook folder for ts Transpile
