+++ typescript/packages/well-log-viewer/tsconfig.json
@@ -8,9 +8,9 @@
     "exclude": [
         "src/demo",
         "src/test/",
+        "src/Storybook/",
         "./**/*.test.tsx",
         "./**/*.stories.tsx",
         "./**/*.stories.mdx",
-        "./**/*.story-example.*"
     ]
 }
